language/cc: add fileInfo (#121)
* language/cc: replace sourceFile type with string

sourceFile was a pretty thin abstraction: we frequently needed to
convert it to and from string. This prepares us to introduce
a fileInfo type that contains more information.

For #73

Signed-off-by: Jay Conrod <jay@engflow.com>

* language/cc: add fileInfo

fileInfo contains metadata describing a file. It's primarily useful
to track a file's kind (libHdrKind, testSrcKind, etc) and includes.

In the next PR, fileInfo will replace ccSourceInfoSet.

For #73

Signed-off-by: Jay Conrod <jay@engflow.com>

---------

Signed-off-by: Jay Conrod <jay@engflow.com>