language/cc: replace ccSourceInfoSet with list of fileInfos (#122)
* language/cc: replace sourceFile type with string

sourceFile was a pretty thin abstraction: we frequently needed to
convert it to and from string. This prepares us to introduce
a fileInfo type that contains more information.

For #73

Signed-off-by: Jay Conrod <jay@engflow.com>

* language/cc: add fileInfo

fileInfo contains metadata describing a file. It's primarily useful
to track a file's kind (libHdrKind, testSrcKind, etc) and includes.

In the next PR, fileInfo will replace ccSourceInfoSet.

For #73

Signed-off-by: Jay Conrod <jay@engflow.com>

* language/cc: replace ccSourceInfoSet with list of fileInfos

ccSourceInfoSet was a struct of lists, classifying file names by kind.
This change replaces it with a list of fileInfo structs, embedding
their own classification.

This refactoring is intended to support a future "subdirectory" mode,
where files are classified based on their containing directory name,
not only by file extension. We flatten files with multiple extensions
into lists in several places, so we lose classification information
if they're lists of strings. Storing each files "kind" alongside its
name makes this change much easier.

Also deleted the sourceInfos map type, since it's no longer needed.

For #73

Signed-off-by: Jay Conrod <jay@engflow.com>

* Apply suggestion from @WojciechMazur

Co-authored-by: Wojciech Mazur <wmazur@virtuslab.com>
Signed-off-by: Jay Conrod <jay@engflow.com>

* split condition in generateLibraryRules

Signed-off-by: Jay Conrod <jay@engflow.com>

---------

Signed-off-by: Jay Conrod <jay@engflow.com>
Co-authored-by: Wojciech Mazur <wmazur@virtuslab.com>