Release presubmit: upgrade example googletest, test in GitHub Actions (#153)
Added a GitHub Actions job that emulates .bcr/presubmit.yml, the
post-release test that BCR runs. We can't run that test directly,
but we should be confident that it's going to pass. (For some reason,
I thought we were already testing this. Guess not.)

In example/bzlmod, upgraded googletest from 1.17.0 to 1.17.0.bcr.2.
The old version does not explicitly load "cc_library" and is not
compatible with Bazel 9.

Signed-off-by: Jay Conrod <jay@engflow.com>