Do not use pointer receiver for Set methods
In Go, maps are reference types and are passed by reference
semantically. Updates made inside a function are visible in the outer
scope, and there's no need to explicitly pass a pointer to the map. At
some specific contexts Go does not allow calling methods with pointer
receivers on map types, so it's better to avoid pointer receivers.

Signed-off-by: Mateusz Olczyk <molczyk@virtuslab.com>